feat: move antarctica to asia
