+++ docs/assets/Routes-Cfw1XFYV.js
@@ -1,0 +1,3 @@
+const __vite__mapDeps=(i,m=__vite__mapDeps,d=(m.f||(m.f=["assets/cars-BzJ-Tnyh.js","assets/rolldown-runtime-DF2fYuay.js","assets/react-vendor-HqBF-KSi.js","assets/charts-DOPwoKjG.js","assets/createSimplePaletteValueFilter-C4uqQZG8.js","assets/Button-DPflrBYw.js","assets/ButtonBase-TQi9tLMq.js","assets/useEnhancedEffect-TwgxY3bQ.js","assets/CircularProgress-BVBDdCOb.js","assets/useId-BseCNtMg.js","assets/useId-DzlSYYoU.js","assets/ButtonGroup-D7RYNrtW.js","assets/Card-BDl1g5Lx.js","assets/CardContent-BMCn_NGV.js","assets/Grid-B41XEp5L.js","assets/useThemeProps-CE8g2MKl.js","assets/isMuiElement-DqsVGliM.js","assets/Typography-dtZqBMwM.js","assets/amber-BE5TiSxW.js","assets/indigo-DB5IViI9.js","assets/teal-Dxj8fkd6.js","assets/yellow-DQPfIU75.js","assets/react-CuPne-I1.js","assets/DateHelper-Bwgbx9Lu.js","assets/expandable-card-C4SsMyX6.js","assets/CardHeader-3yTLPU0x.js","assets/useSlot-D4zBCvp5.js","assets/Collapse-DuQG9KVF.js","assets/utils-BvRsBlDd.js","assets/Row-B_jMErIw.js","assets/theme-atom-BK4g02v0.js","assets/comp-calc-BUZkWyIa.js","assets/Chip-Bf-EVf0P.js","assets/createSvgIcon-BBRGa9WV.js","assets/DialogTitle-3-1fDA0x.js","assets/SelectFocusSourceContext-HarVQ4pI.js","assets/getActiveElement-C1xZk0Mv.js","assets/Divider-DRIqDUkb.js","assets/MenuItem-BK_2bJjz.js","assets/Select-BHBLft5q.js","assets/useControlled-4p42-1Lt.js","assets/RtlProvider-DuGPPv0j.js","assets/useSlotProps-DUsTAVFW.js","assets/TextField-BjtaNL8I.js","assets/lightGreen-ChClOSct.js","assets/education-DVMO4ipt.js","assets/Degree-C-qX3jnZ.js","assets/classes-ZdQfYYOe.js","assets/pink-CLFvxCBy.js","assets/f1-BU_dQ9KF.js","assets/IconButton-BhtOteSy.js","assets/presentations-DVqxqhyM.js","assets/resume-Bci3z9Vp.js","assets/summary-zkSUCAD9.js","assets/Link-DOt6XuAD.js","assets/TableRow-BpKOahMX.js","assets/Job-8yyFy8Mx.js","assets/blueGrey-CmUx3e8f.js","assets/cyan-sGMB-DVK.js","assets/travel-map-BqSHF6GB.js","assets/TableHead-CU9UzqJ0.js","assets/work-C4bYLwbe.js"])))=>i.map(i=>d[i]);
+import{n as e}from"./rolldown-runtime-DF2fYuay.js";import{_ as t,b as n,g as r,h as i,x as a,z as o}from"./react-vendor-HqBF-KSi.js";import{r as s}from"./SelectFocusSourceContext-HarVQ4pI.js";import{t as c}from"./createSvgIcon-BBRGa9WV.js";import{t as l}from"./Header-vuPqoWJM.js";import{t as u}from"./MenuItem-BK_2bJjz.js";import{t as d}from"./Divider-DRIqDUkb.js";import{t as f}from"./loading-spinner-CA3JUfEQ.js";var p=e(o()),m=a(),h=c((0,m.jsx)(`path`,{d:`M12 1.27a11 11 0 00-3.48 21.46c.55.09.73-.28.73-.55v-1.84c-3.03.64-3.67-1.46-3.67-1.46-.55-1.29-1.28-1.65-1.28-1.65-.92-.65.1-.65.1-.65 1.1 0 1.73 1.1 1.73 1.1.92 1.65 2.57 1.2 3.21.92a2 2 0 01.64-1.47c-2.47-.27-5.04-1.19-5.04-5.5 0-1.1.46-2.1 1.2-2.84a3.76 3.76 0 010-2.93s.91-.28 3.11 1.1c1.8-.49 3.7-.49 5.5 0 2.1-1.38 3.02-1.1 3.02-1.1a3.76 3.76 0 010 2.93c.83.74 1.2 1.74 1.2 2.94 0 4.21-2.57 5.13-5.04 5.4.45.37.82.92.82 2.02v3.03c0 .27.1.64.73.55A11 11 0 0012 1.27`}),`GitHub`),g=c((0,m.jsx)(`path`,{d:`M19 3a2 2 0 0 1 2 2v14a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V5a2 2 0 0 1 2-2h14m-.5 15.5v-5.3a3.26 3.26 0 0 0-3.26-3.26c-.85 0-1.84.52-2.32 1.3v-1.11h-2.79v8.37h2.79v-4.93c0-.77.62-1.4 1.39-1.4a1.4 1.4 0 0 1 1.4 1.4v4.93h2.79M6.88 8.56a1.68 1.68 0 0 0 1.68-1.68c0-.93-.75-1.69-1.68-1.69a1.69 1.69 0 0 0-1.69 1.69c0 .93.76 1.68 1.69 1.68m1.39 9.94v-8.37H5.5v8.37h2.77z`}),`LinkedIn`),_=c((0,m.jsx)(`path`,{d:`M19 19H5V5h7V3H5c-1.11 0-2 .9-2 2v14c0 1.1.89 2 2 2h14c1.1 0 2-.9 2-2v-7h-2zM14 3v2h3.59l-9.83 9.83 1.41 1.41L19 6.41V10h2V3z`}),`OpenInNew`),v=()=>{window.open(`https://github.com/kennethbigler/react-home`)},y=()=>{window.open(`https://www.linkedin.com/in/kennethbigler`)},b=(0,p.memo)(({onItemClick:e})=>(0,m.jsxs)(s,{children:[(0,p.useMemo)(()=>[{name:`Summary`,route:``},{name:`Work`,route:`work`},{name:`Resume`,route:`resume`},{name:`Education`,route:`education`},{name:`Presentations`,route:`presentations`},{name:`Comp Calculator`,route:`comp`},{divider:!0},{name:`F1`,route:`f1`},{name:`Cars`,route:`cars`},{name:`Travel Map`,route:`travel`},{divider:!0},{name:`Games`,route:`games`}].map((t,n)=>t.divider?(0,m.jsx)(d,{"aria-hidden":!0},n):(0,m.jsx)(u,{onClick:()=>e&&e(`/${t.route||``}`),children:t.name},t.name)),[e]),(0,m.jsx)(d,{"aria-hidden":!0}),(0,m.jsxs)(u,{onClick:v,"aria-label":`GitHub (opens in new tab)`,children:[(0,m.jsx)(h,{"aria-hidden":!0}),`\xA0GitHub\xA0`,(0,m.jsx)(_,{fontSize:`small`,"aria-hidden":!0})]}),(0,m.jsxs)(u,{onClick:y,"aria-label":`LinkedIn (opens in new tab)`,children:[(0,m.jsx)(g,{"aria-hidden":!0}),`\xA0LinkedIn\xA0`,(0,m.jsx)(_,{fontSize:`small`,"aria-hidden":!0})]})]}));b.displayName=`Menu`;var x=(0,p.lazy)(()=>n(()=>import(`./cars-BzJ-Tnyh.js`),__vite__mapDeps([0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30]))),S=(0,p.lazy)(()=>n(()=>import(`./comp-calc-BUZkWyIa.js`),__vite__mapDeps([31,1,2,3,4,5,6,7,8,9,10,12,25,17,26,13,32,33,34,35,28,36,37,38,39,40,41,16,42,43,14,15,44,22,23,30,19,20]))),C=(0,p.lazy)(()=>n(()=>import(`./education-DVMO4ipt.js`),__vite__mapDeps([45,1,2,3,17,4,46,14,15,16,24,12,6,7,25,26,13,27,28,47,19,48,23]))),w=(0,p.lazy)(()=>n(()=>import(`./f1-BU_dQ9KF.js`),__vite__mapDeps([49,1,2,3,4,14,15,16,50,6,7,8,9,10,17,44,22,23,24,12,25,26,13,27,28,29,30,19,20]))),T=(0,p.lazy)(()=>n(()=>import(`./presentations-DVqxqhyM.js`),__vite__mapDeps([51,1,2,3,17,4,46,14,15,16,24,12,6,7,25,26,13,27,28,47,19,48,23]))),E=(0,p.lazy)(()=>n(()=>import(`./resume-Bci3z9Vp.js`),__vite__mapDeps([52,1,2,3,17,4]))),D=(0,p.lazy)(()=>n(()=>import(`./summary-zkSUCAD9.js`),__vite__mapDeps([53,1,2,3,14,4,15,16,54,6,7,17,55,24,12,25,26,13,27,28,56,32,33,57,58,19,21,47,48,23]))),O=(0,p.lazy)(()=>n(()=>import(`./travel-map-BqSHF6GB.js`),__vite__mapDeps([59,1,2,3,4,14,15,16,55,60,17,18,57,58,19,21,22,23,24,12,6,7,25,26,13,27,28,30,20]))),k=(0,p.lazy)(()=>n(()=>import(`./work-C4bYLwbe.js`),__vite__mapDeps([61,1,2,3,14,4,15,16,17,56,32,6,7,33,26,57,58,19,21,24,12,25,13,27,28]))),A=({handleNav:e})=>(0,m.jsxs)(m.Fragment,{children:[(0,m.jsx)(l,{handleNav:e,children:e=>(0,m.jsx)(b,{onItemClick:e})}),(0,m.jsx)(p.Suspense,{fallback:(0,m.jsx)(f,{}),children:(0,m.jsxs)(t,{children:[(0,m.jsx)(r,{path:``,element:(0,m.jsx)(D,{})}),(0,m.jsx)(r,{path:`cars`,element:(0,m.jsx)(x,{})}),(0,m.jsx)(r,{path:`comp`,element:(0,m.jsx)(S,{})}),(0,m.jsx)(r,{path:`education`,element:(0,m.jsx)(C,{})}),(0,m.jsx)(r,{path:`f1`,element:(0,m.jsx)(w,{})}),(0,m.jsx)(r,{path:`presentations`,element:(0,m.jsx)(T,{})}),(0,m.jsx)(r,{path:`resume`,element:(0,m.jsx)(E,{})}),(0,m.jsx)(r,{path:`travel`,element:(0,m.jsx)(O,{})}),(0,m.jsx)(r,{path:`work`,element:(0,m.jsx)(k,{})}),(0,m.jsx)(r,{path:`*`,element:(0,m.jsx)(i,{to:`/`})})]})})]});export{A as default};
+//# sourceMappingURL=Routes-Cfw1XFYV.js.map
