--- docs/assets/index-DBeL36ud.js
@@ -1,30 +1,0 @@
-const __vite__mapDeps=(i,m=__vite__mapDeps,d=(m.f||(m.f=["assets/Routes-D5Pgkp4X.js","assets/rolldown-runtime-DF2fYuay.js","assets/react-vendor-HqBF-KSi.js","assets/charts-DOPwoKjG.js","assets/Header-vuPqoWJM.js","assets/ButtonBase-TQi9tLMq.js","assets/createSimplePaletteValueFilter-C4uqQZG8.js","assets/useEnhancedEffect-TwgxY3bQ.js","assets/SelectFocusSourceContext-HarVQ4pI.js","assets/utils-BvRsBlDd.js","assets/getActiveElement-C1xZk0Mv.js","assets/useSlot-D4zBCvp5.js","assets/IconButton-BhtOteSy.js","assets/CircularProgress-BVBDdCOb.js","assets/useId-BseCNtMg.js","assets/useId-DzlSYYoU.js","assets/createSvgIcon-BBRGa9WV.js","assets/Switch-BW5OzMky.js","assets/useControlled-4p42-1Lt.js","assets/Toolbar-CdisT2hO.js","assets/Typography-dtZqBMwM.js","assets/RtlProvider-DuGPPv0j.js","assets/react-CuPne-I1.js","assets/theme-atom-BK4g02v0.js","assets/indigo-DB5IViI9.js","assets/teal-Dxj8fkd6.js","assets/Divider-DRIqDUkb.js","assets/MenuItem-BK_2bJjz.js","assets/loading-spinner-CA3JUfEQ.js","assets/Routes-D54OyU24.js","assets/menu-items-eWoGV10n.js"])))=>i.map(i=>d[i]);
-import{n as e}from"./rolldown-runtime-DF2fYuay.js";import{E as t,F as n,L as r,N as i,T as a,_ as o,b as s,g as c,m as l,v as u,x as d,y as f,z as p}from"./react-vendor-HqBF-KSi.js";import{B as m,D as h,V as g,d as _,g as v,i as y,j as b,k as x,m as S,n as C}from"./createSimplePaletteValueFilter-C4uqQZG8.js";import{t as ee}from"./useEnhancedEffect-TwgxY3bQ.js";import{t as te}from"./RtlProvider-DuGPPv0j.js";import{t as ne}from"./useId-DzlSYYoU.js";import{n as w}from"./react-CuPne-I1.js";import{r as re}from"./theme-atom-BK4g02v0.js";import{t as T}from"./loading-spinner-CA3JUfEQ.js";(function(){let e=document.createElement(`link`).relList;if(e&&e.supports&&e.supports(`modulepreload`))return;for(let e of document.querySelectorAll(`link[rel="modulepreload"]`))n(e);new MutationObserver(e=>{for(let t of e)if(t.type===`childList`)for(let e of t.addedNodes)e.tagName===`LINK`&&e.rel===`modulepreload`&&n(e)}).observe(document,{childList:!0,subtree:!0});function t(e){let t={};return e.integrity&&(t.integrity=e.integrity),e.referrerPolicy&&(t.referrerPolicy=e.referrerPolicy),e.crossOrigin===`use-credentials`?t.credentials=`include`:e.crossOrigin===`anonymous`?t.credentials=`omit`:t.credentials=`same-origin`,t}function n(e){if(e.ep)return;e.ep=!0;let n=t(e);fetch(e.href,n)}})();var E=e(p()),D=d(),O=new Map,k={insert:void 0},A=(e,t)=>{let n=i(e);return n.sheet=new t({key:n.key,nonce:n.sheet.nonce,container:n.sheet.container,speedy:n.sheet.isSpeedy,prepend:n.sheet.prepend,insertionPoint:n.sheet.insertionPoint}),n},j;if(typeof document==`object`&&(j=document.querySelector(`[name="emotion-insertion-point"]`),!j)){j=document.createElement(`meta`),j.setAttribute(`name`,`emotion-insertion-point`),j.setAttribute(`content`,``);let e=document.querySelector(`head`);e&&e.prepend(j)}function M(e,t){if(e||t){class r extends n{insert(e,t){return k.insert?k.insert(e,t):(this.key&&this.key.endsWith(`global`)&&(this.before=j),super.insert(e,t))}}let i=A({key:`css`,insertionPoint:e?j:void 0},r);if(t){let e=i.insert;i.insert=(...t)=>(t[1].styles.match(/^@layer\s+[^{]*$/)||(t[1].styles=`@layer mui {${t[1].styles}}`),e(...t))}return i}}function N(e){let{injectFirst:t,enableCssLayer:n,children:r}=e,i=E.useMemo(()=>{let e=`${t}-${n}`;if(typeof document==`object`&&O.has(e))return O.get(e);let r=M(t,n);return O.set(e,r),r},[t,n]);return i?(0,D.jsx)(a,{value:i,children:r}):r}var P=E.createContext(null);function F(){return E.useContext(P)}var I=typeof Symbol==`function`&&Symbol.for?Symbol.for(`mui.nested`):`__THEME_NESTED__`;function L(e,t){return typeof t==`function`?t(e):{...e,...t}}function R(e){let{children:t,theme:n}=e,r=F(),i=E.useMemo(()=>{let e=r===null?{...n}:L(r,n);return e!=null&&(e[I]=r!==null),e},[n,r]);return(0,D.jsx)(P.Provider,{value:i,children:t})}function ie(e){let t=x(),n=ne()||``,{modularCssLayers:r}=e,i=`mui.global, mui.components, mui.theme, mui.custom, mui.sx`;return i=!r||t!==null?``:typeof r==`string`?r.replace(/mui(?!\.)/g,i):`@layer ${i};`,ee(()=>{let e=document.querySelector(`head`);if(!e)return;let t=e.firstChild;if(i){if(t&&t.hasAttribute?.(`data-mui-layer-order`)&&t.getAttribute(`data-mui-layer-order`)===n)return;let r=document.createElement(`style`);r.setAttribute(`data-mui-layer-order`,n),r.textContent=i,e.prepend(r)}else e.querySelector(`style[data-mui-layer-order="${n}"]`)?.remove()},[i,n]),i?(0,D.jsx)(h,{styles:i}):null}var z={};function B(e,t,n,r=!1){return E.useMemo(()=>{let i=e&&t[e]||t;if(typeof n==`function`){let a=n(i),o=e?{...t,[e]:a}:a;return r?()=>o:o}return e?{...t,[e]:n}:{...t,...n}},[e,t,n,r])}function ae(e){let{children:n,theme:r,themeId:i}=e,a=x(z),o=F()||z,s=B(i,a,r),c=B(i,o,r,!0),l=(i?s[i]:s).direction===`rtl`,u=ie(s);return(0,D.jsx)(R,{theme:c,children:(0,D.jsx)(t.Provider,{value:s,children:(0,D.jsx)(te,{value:l,children:(0,D.jsxs)(v,{value:i?s[i].components:s.components,children:[u,n]})})})})}var V=`mode`,H=`color-scheme`,U=`data-color-scheme`;function oe(e){let{defaultMode:t=`system`,defaultLightColorScheme:n=`light`,defaultDarkColorScheme:r=`dark`,modeStorageKey:i=V,colorSchemeStorageKey:a=H,attribute:o=U,colorSchemeNode:s=`document.documentElement`,nonce:c}=e||{},l=``,u=o;if(o===`class`&&(u=`.%s`),o===`data`&&(u=`[data-%s]`),u.startsWith(`.`)){let e=u.substring(1);l+=`${s}.classList.remove('${e}'.replace('%s', light), '${e}'.replace('%s', dark));
-      ${s}.classList.add('${e}'.replace('%s', colorScheme));`}let d=u.match(/\[([^[\]]+)\]/);if(d){let[e,t]=d[1].split(`=`);t||(l+=`${s}.removeAttribute('${e}'.replace('%s', light));
-      ${s}.removeAttribute('${e}'.replace('%s', dark));`),l+=`
-      ${s}.setAttribute('${e}'.replace('%s', colorScheme), ${t?`${t}.replace('%s', colorScheme)`:`""`});`}else u!==`.%s`&&(l+=`${s}.setAttribute('${u}', colorScheme);`);return(0,D.jsx)(`script`,{suppressHydrationWarning:!0,nonce:typeof window>`u`?c:``,dangerouslySetInnerHTML:{__html:`(function() {
-try {
-  let colorScheme = '';
-  const mode = localStorage.getItem('${i}') || '${t}';
-  const dark = localStorage.getItem('${a}-dark') || '${r}';
-  const light = localStorage.getItem('${a}-light') || '${n}';
-  if (mode === 'system') {
-    // handle system mode
-    const mql = window.matchMedia('(prefers-color-scheme: dark)');
-    if (mql.matches) {
-      colorScheme = dark
-    } else {
-      colorScheme = light
-    }
-  }
-  if (mode === 'light') {
-    colorScheme = light;
-  }
-  if (mode === 'dark') {
-    colorScheme = dark;
-  }
-  if (colorScheme) {
-    ${l}
-  }
-} catch(e){}})();`}},`mui-color-scheme-init`)}function W(){}var G=({key:e,storageWindow:t})=>(!t&&typeof window<`u`&&(t=window),{get(n){if(typeof window>`u`)return;if(!t)return n;let r;try{r=t.localStorage.getItem(e)}catch{}return r||n},set:n=>{if(t)try{t.localStorage.setItem(e,n)}catch{}},subscribe:n=>{if(!t)return W;let r=t=>{let r=t.newValue;t.key===e&&n(r)};return t.addEventListener(`storage`,r),()=>{t.removeEventListener(`storage`,r)}}});function K(){}function q(e){if(typeof window<`u`&&typeof window.matchMedia==`function`&&e===`system`)return window.matchMedia(`(prefers-color-scheme: dark)`).matches?`dark`:`light`}function J(e,t){if(e.mode===`light`||e.mode===`system`&&e.systemMode===`light`)return t(`light`);if(e.mode===`dark`||e.mode===`system`&&e.systemMode===`dark`)return t(`dark`)}function Y(e){return J(e,t=>{if(t===`light`)return e.lightColorScheme;if(t===`dark`)return e.darkColorScheme})}function se(e){let{defaultMode:t=`light`,defaultLightColorScheme:n,defaultDarkColorScheme:r,supportedColorSchemes:i=[],modeStorageKey:a=V,colorSchemeStorageKey:o=H,storageWindow:s=typeof window>`u`?void 0:window,storageManager:c=G,noSsr:l=!1}=e,u=i.join(`,`),d=i.length>1,f=E.useMemo(()=>c?.({key:a,storageWindow:s}),[c,a,s]),p=E.useMemo(()=>c?.({key:`${o}-light`,storageWindow:s}),[c,o,s]),m=E.useMemo(()=>c?.({key:`${o}-dark`,storageWindow:s}),[c,o,s]),[h,g]=E.useState(()=>{let e=f?.get(t)||t,i=p?.get(n)||n,a=m?.get(r)||r;return{mode:e,systemMode:q(e),lightColorScheme:i,darkColorScheme:a}}),[_,v]=E.useState(l||!d);E.useEffect(()=>{v(!0)},[]);let y=Y(h),b=E.useCallback(e=>{g(n=>{if(e===n.mode)return n;let r=e??t;return f?.set(r),{...n,mode:r,systemMode:q(r)}})},[f,t]),x=E.useCallback(e=>{e?typeof e==`string`?e&&!u.includes(e)?console.error(`\`${e}\` does not exist in \`theme.colorSchemes\`.`):g(t=>{let n={...t};return J(t,t=>{t===`light`&&(p?.set(e),n.lightColorScheme=e),t===`dark`&&(m?.set(e),n.darkColorScheme=e)}),n}):g(t=>{let i={...t},a=e.light===null?n:e.light,o=e.dark===null?r:e.dark;return a&&(u.includes(a)?(i.lightColorScheme=a,p?.set(a)):console.error(`\`${a}\` does not exist in \`theme.colorSchemes\`.`)),o&&(u.includes(o)?(i.darkColorScheme=o,m?.set(o)):console.error(`\`${o}\` does not exist in \`theme.colorSchemes\`.`)),i}):g(e=>(p?.set(n),m?.set(r),{...e,lightColorScheme:n,darkColorScheme:r}))},[u,p,m,n,r]),S=E.useCallback(e=>{h.mode===`system`&&g(t=>{let n=e?.matches?`dark`:`light`;return t.systemMode===n?t:{...t,systemMode:n}})},[h.mode]),C=E.useRef(S);return C.current=S,E.useEffect(()=>{if(typeof window.matchMedia!=`function`||!d)return;let e=(...e)=>C.current(...e),t=window.matchMedia(`(prefers-color-scheme: dark)`);return t.addListener(e),e(t),()=>{t.removeListener(e)}},[d]),E.useEffect(()=>{if(d){let e=f?.subscribe(e=>{(!e||[`light`,`dark`,`system`].includes(e))&&b(e||t)})||K,n=p?.subscribe(e=>{(!e||u.match(e))&&x({light:e})})||K,r=m?.subscribe(e=>{(!e||u.match(e))&&x({dark:e})})||K;return()=>{e(),n(),r()}}},[x,b,u,t,s,d,f,p,m]),{...h,mode:_?h.mode:void 0,systemMode:_?h.systemMode:void 0,colorScheme:_?y:void 0,setMode:b,setColorScheme:x}}var ce=`*{-webkit-transition:none!important;-moz-transition:none!important;-o-transition:none!important;-ms-transition:none!important;transition:none!important}`;function le(e){let{themeId:t,theme:n={},modeStorageKey:r=V,colorSchemeStorageKey:i=H,disableTransitionOnChange:a=!1,defaultColorScheme:o,resolveTheme:s}=e,c={allColorSchemes:[],colorScheme:void 0,darkColorScheme:void 0,lightColorScheme:void 0,mode:void 0,setColorScheme:()=>{},setMode:()=>{},systemMode:void 0},l=E.createContext(void 0),u=()=>E.useContext(l)||c,d={},f={};function p(e){let{children:c,theme:u,modeStorageKey:p=r,colorSchemeStorageKey:h=i,disableTransitionOnChange:g=a,storageManager:_,storageWindow:v=typeof window>`u`?void 0:window,documentNode:y=typeof document>`u`?void 0:document,colorSchemeNode:b=typeof document>`u`?void 0:document.documentElement,disableNestedContext:x=!1,disableStyleSheetGeneration:S=!1,defaultMode:C=`system`,forceThemeRerender:te=!1,noSsr:ne}=e,w=E.useRef(!1),re=F(),T=E.useContext(l),O=!!T&&!x,k=E.useMemo(()=>u||(typeof n==`function`?n():n),[u]),A=k[t],j=A||k,{colorSchemes:M=d,components:N=f,cssVarPrefix:P}=j,I=Object.keys(M).filter(e=>!!M[e]).join(`,`),L=E.useMemo(()=>I.split(`,`),[I]),R=typeof o==`string`?o:o.light,ie=typeof o==`string`?o:o.dark,{mode:z,setMode:B,systemMode:V,lightColorScheme:H,darkColorScheme:U,colorScheme:oe,setColorScheme:W}=se({supportedColorSchemes:L,defaultLightColorScheme:R,defaultDarkColorScheme:ie,modeStorageKey:p,colorSchemeStorageKey:h,defaultMode:M[R]&&M[ie]?C:M[j.defaultColorScheme]?.palette?.mode||j.palette?.mode,storageManager:_,storageWindow:v,noSsr:ne}),G=z,K=oe;O&&(G=T.mode,K=T.colorScheme);let q=K||j.defaultColorScheme;j.vars&&!te&&(q=j.defaultColorScheme);let J=E.useMemo(()=>{let e=j.generateThemeVars?.()||j.vars,t={...j,components:N,colorSchemes:M,cssVarPrefix:P,vars:e};if(typeof t.generateSpacing==`function`&&(t.spacing=t.generateSpacing()),q){let e=M[q];e&&typeof e==`object`&&Object.keys(e).forEach(n=>{e[n]&&typeof e[n]==`object`?t[n]={...t[n],...e[n]}:t[n]=e[n]})}return s?s(t):t},[j,q,N,M,P]),Y=j.colorSchemeSelector;ee(()=>{if(K&&b&&Y&&Y!==`media`){let e=Y,t=Y;if(e===`class`&&(t=`.%s`),e===`data`&&(t=`[data-%s]`),e?.startsWith(`data-`)&&!e.includes(`%s`)&&(t=`[${e}="%s"]`),t.startsWith(`.`))b.classList.remove(...L.map(e=>t.substring(1).replace(`%s`,e))),b.classList.add(t.substring(1).replace(`%s`,K));else{let e=t.replace(`%s`,K).match(/\[([^\]]+)\]/);if(e){let[t,n]=e[1].split(`=`);n||L.forEach(e=>{b.removeAttribute(t.replace(K,e))}),b.setAttribute(t,n?n.replace(/"|'/g,``):``)}else b.setAttribute(t,K)}}},[K,Y,b,L]),E.useEffect(()=>{let e;if(g&&w.current&&y){let t=y.createElement(`style`);t.appendChild(y.createTextNode(ce)),y.head.appendChild(t),window.getComputedStyle(y.body),e=setTimeout(()=>{y.head.removeChild(t)},1)}return()=>{clearTimeout(e)}},[K,g,y]),E.useEffect(()=>(w.current=!0,()=>{w.current=!1}),[]);let le=E.useMemo(()=>({allColorSchemes:L,colorScheme:K,darkColorScheme:U,lightColorScheme:H,mode:G,setColorScheme:W,setMode:B,systemMode:V}),[L,K,U,H,G,W,B,V,J.colorSchemeSelector]),X=!0;(S||j.cssVariables===!1||O&&re?.cssVarPrefix===P)&&(X=!1);let Z=(0,D.jsxs)(E.Fragment,{children:[(0,D.jsx)(ae,{themeId:A?t:void 0,theme:J,children:c}),X&&(0,D.jsx)(m,{styles:J.generateStyleSheets?.()||[]})]});return O?Z:(0,D.jsx)(l.Provider,{value:le,children:Z})}let h=typeof o==`string`?o:o.light,g=typeof o==`string`?o:o.dark;return{CssVarsProvider:p,useColorScheme:u,getInitColorSchemeScript:e=>oe({colorSchemeStorageKey:i,defaultLightColorScheme:h,defaultDarkColorScheme:g,modeStorageKey:r,...e})}}function X({theme:e,...t}){let n=`$$material`in e?e[g]:void 0;return(0,D.jsx)(ae,{...t,themeId:n?g:void 0,theme:n||e})}var Z={attribute:`data-mui-color-scheme`,colorSchemeStorageKey:`mui-color-scheme`,defaultLightColorScheme:`light`,defaultDarkColorScheme:`dark`,modeStorageKey:`mui-mode`},{CssVarsProvider:ue,useColorScheme:de,getInitColorSchemeScript:fe}=le({themeId:g,theme:()=>_({cssVariables:!0}),colorSchemeStorageKey:Z.colorSchemeStorageKey,modeStorageKey:Z.modeStorageKey,defaultColorScheme:{light:Z.defaultLightColorScheme,dark:Z.defaultDarkColorScheme},resolveTheme:e=>{let t={...e,typography:S(e.palette,e.typography)};return t.unstable_sx=function(e){return b({sx:e,theme:this})},t}}),pe=ue;function me({theme:e,...t}){let n=E.useMemo(()=>{if(typeof e==`function`)return e;let t=`$$material`in e?e[g]:e;return`colorSchemes`in t?null:`vars`in t?e:{...e,vars:null}},[e]);return n?(0,D.jsx)(X,{theme:n,...t}):(0,D.jsx)(pe,{theme:e,...t})}var Q=typeof y({})==`function`,he=(e,t)=>({WebkitFontSmoothing:`antialiased`,MozOsxFontSmoothing:`grayscale`,boxSizing:`border-box`,WebkitTextSizeAdjust:`100%`,...t&&!e.vars&&{colorScheme:e.palette.mode}}),ge=e=>({color:(e.vars||e).palette.text.primary,...e.typography.body1,backgroundColor:(e.vars||e).palette.background.default,"@media print":{backgroundColor:(e.vars||e).palette.common.white}}),_e=(e,t=!1)=>{let n={};t&&e.colorSchemes&&typeof e.getColorSchemeSelector==`function`&&Object.entries(e.colorSchemes).forEach(([t,r])=>{let i=e.getColorSchemeSelector(t);i.startsWith(`@`)?n[i]={":root":{colorScheme:r.palette?.mode}}:n[i.replace(/\s*&/,``)]={colorScheme:r.palette?.mode}});let r={html:he(e,t),"*, *::before, *::after":{boxSizing:`inherit`},"strong, b":{fontWeight:e.typography.fontWeightBold},body:{margin:0,...ge(e),"&::backdrop":{backgroundColor:(e.vars||e).palette.background.default}},...n},i=e.components?.MuiCssBaseline?.styleOverrides;return i&&(r=[r,i]),r},$=`mui-ecs`,ve=e=>{let t=_e(e,!1),n=Array.isArray(t)?t[0]:t;return!e.vars&&n&&(n.html[`:root:has(${$})`]={colorScheme:e.palette.mode}),e.colorSchemes&&Object.entries(e.colorSchemes).forEach(([t,r])=>{let i=e.getColorSchemeSelector(t);i.startsWith(`@`)?n[i]={[`:root:not(:has(.${$}))`]:{colorScheme:r.palette?.mode}}:n[i.replace(/\s*&/,``)]={[`&:not(:has(.${$}))`]:{colorScheme:r.palette?.mode}}}),t},ye=y(Q?({theme:e,enableColorScheme:t})=>_e(e,t):({theme:e})=>ve(e));function be(e){let{children:t,enableColorScheme:n=!1}=C({props:e,name:`MuiCssBaseline`});return(0,D.jsxs)(E.Fragment,{children:[Q&&(0,D.jsx)(ye,{enableColorScheme:n}),!Q&&!n&&(0,D.jsx)(`span`,{className:$,style:{display:`none`}}),t]})}var xe=e(r(),1),Se=`Ken Bigler's Website`,Ce=e=>e.replace(/\w\S*/g,e=>e.charAt(0).toUpperCase()+e.substring(1).toLowerCase()),we=e=>{let t=e.replace(/^#\/?/,``).split(`/`).filter(Boolean)[0],n=e.replace(/^#\/?/,``).split(`/`).filter(Boolean)[1],r;return r=t?t===`games`?n?`${Ce(n.replace(/-/g,` `))} | Game`:`Games`:Ce(t.replace(/-/g,` `)):Se,r===`Ken Bigler's Website`?Se:`${r} | ${Se}`},Te=(0,E.lazy)(()=>s(()=>import(`./Routes-D5Pgkp4X.js`),__vite__mapDeps([0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28]))),Ee=(0,E.lazy)(()=>s(()=>import(`./Routes-D54OyU24.js`),__vite__mapDeps([29,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,30]))),De=()=>{let e=f(),t=u();(0,E.useEffect)(()=>{let e=t.pathname||`/`;document.title=we(e)},[t.pathname]);let n=t=>{e(t)};return(0,D.jsx)(`main`,{style:{padding:`1em`,paddingTop:`5em`},children:(0,D.jsx)(E.Suspense,{fallback:(0,D.jsx)(T,{}),children:(0,D.jsxs)(o,{children:[(0,D.jsx)(c,{path:`games/*`,element:(0,D.jsx)(Ee,{handleNav:n})}),(0,D.jsx)(c,{path:`/*`,element:(0,D.jsx)(Te,{handleNav:n})})]})})})},Oe=()=>(0,D.jsx)(l,{children:(0,D.jsx)(De,{})}),ke=()=>{let e=w(re);return(0,D.jsx)(N,{injectFirst:!0,children:(0,D.jsxs)(me,{theme:(0,E.useMemo)(()=>_({palette:{mode:e.mode,primary:e.primary,secondary:e.secondary,contrastThreshold:4.5},typography:{fontFamily:[`Montserrat`,`sans-serif`].join(`,`)},breakpoints:{values:{xs:0,sm:600,md:900,lg:1200,xl:1536,xxl:2600,xxxl:4600}}}),[e.mode,e.primary,e.secondary]),children:[(0,D.jsx)(be,{}),(0,D.jsx)(Oe,{})]})})},Ae=document.getElementById(`root`);if(!Ae)throw Error(`Root element #root not found in document.`);xe.createRoot(Ae).render((0,D.jsx)(E.StrictMode,{children:(0,D.jsx)(ke,{})}));
-//# sourceMappingURL=index-DBeL36ud.js.map
