--- docs/assets/travel-map-CV-__66Y.js
@@ -1,2 +1,0 @@
-import{n as e}from"./rolldown-runtime-DF2fYuay.js";import{a as t,c as n,d as r,f as i,i as ee,l as a,n as te,o as ne,p as re,r as o,s as ie,t as ae,u as s,x as oe,z as se}from"./react-vendor-HqBF-KSi.js";import{G as c,U as l,Y as u,q as d}from"./createSimplePaletteValueFilter-C4uqQZG8.js";import{t as f}from"./indigo-DB5IViI9.js";import{t as ce}from"./cyan-sGMB-DVK.js";import{t as p}from"./yellow-DQPfIU75.js";import{t as le}from"./amber-BE5TiSxW.js";import{t as ue}from"./blueGrey-CmUx3e8f.js";import{t as m}from"./Grid-B41XEp5L.js";import{t as h}from"./Typography-dtZqBMwM.js";import{a as g,n as _,r as v,t as y}from"./TableRow-BpKOahMX.js";import{t as de}from"./TableHead-CU9UzqJ0.js";import{n as b}from"./react-CuPne-I1.js";import{r as x}from"./theme-atom-BK4g02v0.js";import{a as fe,i as pe,o as me,r as he,t as ge}from"./charts-DOPwoKjG.js";import{t as S}from"./expandable-card-C4SsMyX6.js";import{t as C}from"./DateHelper-Bwgbx9Lu.js";var w=e(se(),1),T=[{name:`American Samoa`,continent:`AS`,flag:`🇦🇸`,color:c[500]},{name:`Antarctica`,continent:`SA`,flag:`🇦🇶`,color:c[500]},{name:`Argentina`,continent:`SA`,flag:`🇦🇷`,color:c[500]},{name:`Australia`,continent:`AU`,flag:`🇦🇺`,color:c[500]},{name:`Austria`,continent:`EU`,flag:`🇦🇹`,color:p[500]},{name:`Bahamas`,continent:`NA`,flag:`🇧🇸`,color:c[500]},{name:`Brazil`,continent:`SA`,flag:`🇧🇷`,color:p[500]},{name:`British Virgin Islands`,continent:`NA`,flag:`🇻🇬`,color:c[500]},{name:`Canada`,continent:`NA`,flag:`🇨🇦`,color:c[500]},{name:`Cayman Islands`,continent:`NA`,flag:`🇰🇾`,color:c[500]},{name:`Chile`,continent:`SA`,flag:`🇨🇱`,color:p[500]},{name:`China`,continent:`AS`,flag:`🇨🇳`,color:c[500]},{name:`Colombia`,continent:`SA`,flag:`🇨🇴`,color:c[500]},{name:`Denmark`,continent:`EU`,flag:`🇩🇰`,color:c[500]},{name:`Egypt`,continent:`AF`,flag:`🇪🇬`,color:c[500]},{name:`Estonia`,continent:`EU`,flag:`🇪🇪`,color:c[500]},{name:`Fiji`,continent:`AS`,flag:`🇫🇯`,color:c[500]},{name:`Finland`,continent:`EU`,flag:`🇫🇮`,color:c[500]},{name:`France`,continent:`EU`,flag:`🇫🇷`,color:c[500]},{name:`Germany`,continent:`EU`,flag:`🇩🇪`,color:c[500]},{name:`Gibraltar`,continent:`EU`,flag:`🇬🇮`,color:c[500]},{name:`Greece`,continent:`EU`,flag:`🇬🇷`,color:c[500]},{name:`Hong Kong`,continent:`AS`,flag:`🇭🇰`,color:c[500]},{name:`Iceland`,continent:`EU`,flag:`🇮🇸`,color:c[500]},{name:`India`,continent:`AS`,flag:`🇮🇳`,color:c[500]},{name:`Ireland`,continent:`EU`,flag:`🇮🇪`,color:c[500]},{name:`Italy`,continent:`EU`,flag:`🇮🇹`,color:c[500]},{name:`Jamaica`,continent:`NA`,flag:`🇯🇲`,color:c[500]},{name:`Japan`,continent:`AS`,flag:`🇯🇵`,color:c[500]},{name:`Malta`,continent:`EU`,flag:`🇲🇹`,color:c[500]},{name:`Mexico`,continent:`NA`,flag:`🇲🇽`,color:c[500]},{name:`Monaco`,continent:`EU`,flag:`🇲🇨`,color:c[500]},{name:`Morocco`,continent:`AF`,flag:`🇲🇦`,color:c[500]},{name:`Netherlands`,continent:`EU`,flag:`🇳🇱`,color:c[500]},{name:`New Caledonia`,continent:`AS`,flag:`🇳🇨`,color:c[500]},{name:`Norway`,continent:`EU`,flag:`🇳🇴`,color:c[500]},{name:`Panama`,continent:`NA`,flag:`🇵🇦`,color:p[500]},{name:`Poland`,continent:`EU`,flag:`🇵🇱`,color:c[500]},{name:`Portugal`,continent:`EU`,flag:`🇵🇹`,color:c[500]},{name:`Puerto Rico`,continent:`NA`,flag:`🇵🇷`,color:c[500]},{name:`Russia`,continent:`EU`,flag:`🇷🇺`,color:c[500]},{name:`Spain`,continent:`EU`,flag:`🇪🇸`,color:c[500]},{name:`Sweden`,continent:`EU`,flag:`🇸🇪`,color:c[500]},{name:`Switzerland`,continent:`EU`,flag:`🇨🇭`,color:c[500]},{name:`Turkey`,continent:`EU`,flag:`🇹🇷`,color:c[500]},{name:`United Arab Emirates`,continent:`AF`,flag:`🇦🇪`,color:p[500]},{name:`United Kingdom`,continent:`EU`,flag:`🇬🇧`,color:c[500]},{name:`United States of America`,continent:`NA`,flag:`🇺🇸`,color:c[500]},{name:`U.S. Virgin Islands`,continent:`NA`,flag:`🇻🇮`,color:c[500]},{name:`Vatican`,continent:`EU`,flag:`🇻🇦`,color:c[500]}],E=[],D=[],O=[];T.forEach(e=>{switch(e.continent){case`NA`:case`SA`:E.push(e);break;case`EU`:case`AF`:D.push(e);break;default:O.push(e)}});var _e=T.length,k=[{year:2020,vacation:4,work:0},{year:2021,vacation:47,work:0},{year:2022,vacation:41,work:33},{year:2023,vacation:66,work:11},{year:2024,vacation:84,work:23},{year:2025,vacation:78,work:34},{year:2026,vacation:33,work:7}],ve=k.map(e=>e.vacation),ye=k.map(e=>e.work),A=oe(),j={borderRight:`1px solid ${l[400]}`},M={padding:`5px 1px`,textAlign:`center`,whiteSpace:`normal`,overflow:`visible`},N=2,P=[],be=Math.max(E.length,Math.ceil(D.length/N),O.length);for(let e=0;e<be;e+=1){let t=[];t.push((0,A.jsxs)(_,{style:{...M,...j},children:[E[e]?.name||``,` `,E[e]&&(0,A.jsx)(`span`,{"aria-hidden":!0,children:E[e].flag})]},`tmc${e}-americas`));for(let n=0;n<N;n+=1)t.push((0,A.jsxs)(_,{style:n===N-1?{...M,...j}:M,children:[D[N*e+n]?.name||``,` `,D[N*e+n]&&(0,A.jsx)(`span`,{"aria-hidden":!0,children:D[N*e+n].flag})]},`tmc${e}${n}-eu-af`));t.push((0,A.jsxs)(_,{style:M,children:[O[e]?.name||``,` `,O[e]&&(0,A.jsx)(`span`,{"aria-hidden":!0,children:O[e].flag})]},`tmc${e}-as-au`));let n=(0,A.jsx)(y,{children:t},`tmr${e}`);P.push(n)}var F=(0,w.memo)(()=>(0,A.jsxs)(g,{"aria-label":`Countries I have visited`,children:[(0,A.jsx)(de,{children:(0,A.jsxs)(y,{children:[(0,A.jsx)(_,{style:{...M,...j},children:(0,A.jsx)(h,{variant:`h6`,children:`The Americas`})}),(0,A.jsx)(_,{colSpan:N,style:{...M,...j},children:(0,A.jsx)(h,{variant:`h6`,children:`Europe & Africa`})}),(0,A.jsx)(_,{style:M,children:(0,A.jsx)(h,{variant:`h6`,children:`Asia & Australia`})})]})}),(0,A.jsx)(v,{children:P})]}));F.displayName=`CountryTable`;var I=e(ge(),1);fe(),re(I.default);var L={chart:{backgroundColor:`transparent`,height:`60%`}},xe={type:`map`,name:`Visited`,states:{hover:{color:d[500]}},joinBy:[`name`,`name`],showInLegend:!1,tooltip:{pointFormat:`{point.name}: {point.flag}`}},R=(0,w.memo)(()=>{let[e,t]=(0,w.useState)(),[n,r]=(0,w.useState)(!1),i=b(x).mode===`light`?`black`:`white`;(0,w.useEffect)(()=>{let e=new AbortController;return fetch(`https://unpkg.com/world-atlas@2.0.2/countries-110m.json`,{signal:e.signal}).then(e=>e.json()).then(e=>t(e)).catch(e=>{e instanceof Error&&e.name!==`AbortError`&&r(!0)}),()=>e.abort()},[]);let ee=(0,w.useMemo)(()=>({...L,chart:{...L.chart,map:e}}),[e]);return n?(0,A.jsxs)(h,{variant:`h3`,children:[`Map failed to load. `,_e,` Countries Visited.`]}):e?(0,A.jsx)(`figure`,{style:{margin:0,width:`100%`},children:(0,A.jsxs)(ae,{highcharts:I.default,options:ee,children:[(0,A.jsx)(o,{enabled:!0}),(0,A.jsx)(a,{enabled:!1}),(0,A.jsxs)(s,{style:{color:i},children:[`Travel Map: `,_e,` Countries Visited`]}),(0,A.jsx)(te,{type:`map`,options:xe,data:T})]})}):(0,A.jsx)(h,{variant:`h3`,children:`Loading World Map...`})});R.displayName=`WorldMap`;var z=function(e){return e.Disney=`Disney`,e.NG=`National Geographic`,e.Princess=`Princess`,e.RC=`Royal Caribbean`,e.Virgin=`Virgin Voyages`,e}(z||{}),B=function(e){return e.Magic=`Magic`,e.Wonder=`Wonder`,e.Dream=`Dream`,e.Fantasy=`Fantasy`,e.Wish=`Wish`,e.Treasure=`Treasure`,e.Destiny=`Destiny`,e.IcelandAir=`747`,e.Endurance=`Endurance`,e.Discovery=`Discovery`,e.Navigator=`Navigator of the Seas`,e.Scarlet=`Scarlet Lady`,e}(B||{}),V=function(e){return e.Caribbean=`Caribbean`,e.Europe=`Europe`,e.Pacific=`Pacific`,e.West=`America`,e}(V||{}),H=[{departure:C(`2004-04`),line:z.Disney,name:`🇺🇸🇧🇸`,nights:3,ship:B.Wonder,region:V.Caribbean},{departure:C(`2005-06`),line:z.Disney,name:`🇺🇸🇲🇽`,nights:7,ship:B.Magic,region:V.West},{concierge:!0,departure:C(`2011-06`),line:z.Disney,name:`🇪🇸🇲🇹🇮🇹🇫🇷🇲🇨🇻🇦`,nights:11,ship:B.Magic,region:V.Europe},{departure:C(`2012-06`),line:z.Disney,name:`🇺🇸🇨🇦`,nights:7,ship:B.Wonder,region:V.West},{concierge:!0,departure:C(`2013-07`),line:z.Disney,name:`🇪🇸🇫🇷🇲🇨🇮🇹🇻🇦🇬🇷🇹🇷🇲🇹`,nights:12,ship:B.Magic,region:V.Europe},{departure:C(`2014-07`),line:z.Disney,name:`🇺🇸🇨🇦`,nights:7,ship:B.Wonder,region:V.West},{concierge:!0,departure:C(`2015-07`),line:z.Disney,name:`🇬🇧🇩🇰🇸🇪🇫🇮🇷🇺🇪🇪`,nights:14,ship:B.Magic,region:V.Europe},{concierge:!0,departure:C(`2015-08`),line:z.Disney,name:`🇬🇧🇪🇸🇵🇹🇬🇮`,nights:7,ship:B.Magic,region:V.Europe},{departure:C(`2015-12`),line:z.Disney,name:`🇺🇸🇧🇸`,nights:4,ship:B.Dream,region:V.Caribbean},{departure:C(`2016-06`),line:z.Disney,name:`🇬🇧🇮🇸🇳🇴🇩🇰`,nights:12,ship:B.Magic,region:V.Europe},{departure:C(`2016-12`),line:z.Disney,name:`🇺🇸🇻🇬🇻🇮🇧🇸`,nights:7,ship:B.Fantasy,region:V.Caribbean},{departure:C(`2017-05`),line:z.Disney,name:`🇺🇸🇵🇹🇬🇧🇳🇱🇩🇰`,nights:15,ship:B.Magic,region:V.Europe},{departure:C(`2017-12`),line:z.Disney,name:`🇺🇸🇧🇸🇰🇾🇲🇽`,nights:7,ship:B.Fantasy,region:V.Caribbean},{departure:C(`2018-05`),line:z.Disney,name:`🇺🇸🇵🇹🇪🇸`,nights:13,ship:B.Magic,region:V.Europe},{departure:C(`2018-07`),line:z.Disney,name:`🇪🇸🇫🇷🇲🇨🇮🇹🇻🇦`,nights:10,ship:B.Magic,region:V.Europe},{departure:C(`2018-11`),line:z.Disney,name:`🇺🇸🇧🇸`,nights:7,ship:B.Magic,region:V.Caribbean},{concierge:!0,departure:C(`2019-09`),line:z.Disney,name:`🇬🇧🇮🇪🇺🇸`,nights:10,ship:B.Magic,region:V.Europe},{concierge:!0,departure:C(`2019-11`),line:z.Disney,name:`🇺🇸🇲🇽🇰🇾🇯🇲🇧🇸`,nights:7,ship:B.Fantasy,region:V.Caribbean},{concierge:!0,departure:C(`2021-12`),line:z.Disney,name:`🇺🇸🇻🇬🇻🇮🇧🇸`,nights:7,ship:B.Fantasy,region:V.Caribbean},{concierge:!0,departure:C(`2022-04`),line:z.Disney,name:`🇨🇦🇺🇸`,nights:10,ship:B.Wonder,region:V.Pacific},{departure:C(`2022-08`),line:z.RC,name:`🇺🇸🇲🇽`,nights:4,ship:B.Navigator,region:V.West},{departure:C(`2022-12`),line:z.Disney,name:`🇺🇸🇧🇸`,nights:4,ship:B.Wish,region:V.Caribbean},{concierge:!0,departure:C(`2022-12`),line:z.Disney,name:`🇺🇸🇧🇸`,nights:3,ship:B.Wish,region:V.Caribbean},{departure:C(`2023-02`),line:z.Princess,name:`🇺🇸🇲🇽`,nights:4,ship:B.Discovery,region:V.West},{concierge:!0,departure:C(`2023-05`),line:z.Disney,name:`🇺🇸🇧🇸`,nights:3,ship:B.Dream,region:V.Caribbean},{concierge:!0,departure:C(`2023-05`),line:z.Disney,name:`🇺🇸🇵🇹🇪🇸`,nights:13,ship:B.Dream,region:V.Europe},{departure:C(`2023-07`),line:z.NG,name:`🥇-🇺🇸🇯🇵🇨🇳🇭🇰🇮🇳🇪🇬🇫🇷🇨🇦`,nights:23,ship:B.IcelandAir,region:V.Europe},{concierge:!0,departure:C(`2024-03`),line:z.Disney,name:`🇦🇺🇳🇨🇫🇯🇦🇸🇺🇸`,nights:15,ship:B.Wonder,region:V.Pacific},{departure:C(`2024-05`),line:z.Disney,name:`🇺🇸🇧🇸`,nights:3,ship:B.Wish,region:V.Caribbean},{concierge:!0,departure:C(`2024-09`),line:z.Disney,name:`🇨🇦🇺🇸`,nights:10,ship:B.Wonder,region:V.Pacific},{concierge:!0,departure:C(`2024-12`),line:z.Disney,name:`🇺🇸🇧🇸`,nights:4,ship:B.Fantasy,region:V.Caribbean},{concierge:!0,departure:C(`2024-12`),line:z.Disney,name:`🥇-🇺🇸🇻🇬🇻🇮🇧🇸`,nights:7,ship:B.Treasure,region:V.Caribbean},{concierge:!0,departure:C(`2025-03`),line:z.Disney,name:`🇺🇸🇧🇸🇵🇷`,nights:7,ship:B.Magic,region:V.Caribbean},{departure:C(`2025-05`),line:z.Virgin,name:`🇺🇸🇮🇨🇲🇦🇪🇸`,nights:15,ship:B.Scarlet,region:V.Europe},{concierge:!0,departure:C(`2025-10`),line:z.Disney,name:`🇺🇸🇧🇸`,nights:4,ship:B.Dream,region:V.Caribbean},{departure:C(`2025-11`),line:z.Disney,name:`🥇-🇺🇸🇧🇸`,nights:4,ship:B.Destiny,region:V.Caribbean},{departure:C(`2025-12`),line:z.NG,name:`🇦🇷🇨🇱🇦🇶🇧🇷`,nights:11,ship:B.Endurance,region:V.West},{concierge:!0,departure:C(`2026-04`),line:z.Disney,name:`🇺🇸🇲🇽🇨🇴🇵🇦`,nights:14,ship:B.Magic,region:V.West},{concierge:!0,departure:C(`2026-09`),line:z.Disney,name:`🇪🇸🇵🇹🇺🇸`,nights:13,ship:B.Dream,region:V.Europe}],U=[{num:1,status:`Silver`},{num:5,status:`Gold`},{num:10,status:`Platinum`},{num:25,status:`Pearl`}],W=[{num:1,nights:1,status:`Gold`},{num:3,nights:30,status:`Ruby`},{num:5,nights:50,status:`Platinum`},{num:15,nights:150,status:`Elite`}],G=[{nights:3,status:`Gold`},{nights:30,status:`Platinum`},{nights:55,status:`Emerald`},{nights:80,status:`Diamond`},{nights:175,status:`Diamond Plus`},{nights:700,status:`Pinnacle Club`}],Se=[{num:1,status:`Sailing Club`},{num:2,status:`Blue Extras`},{num:4,status:`Deep Blue Extras`}],K=[0,0,0,0],q=[0,0,0,0,0,0,0],J=[0,0],Ce=[0],we=[0],Te=[0],Ee=0,Y=0,De=0,X=[0,0],Oe=0,ke=0,Ae=0,Z=0;H.forEach(e=>{switch(Ee+=e.nights,e.ship){case B.Magic:q[0]+=1;break;case B.Wonder:q[1]+=1;break;case B.Dream:q[2]+=1;break;case B.Fantasy:q[3]+=1;break;case B.Wish:q[4]+=1;break;case B.Treasure:q[5]+=1;break;case B.Destiny:q[6]+=1;break;case B.IcelandAir:J[0]+=1;break;case B.Endurance:J[1]+=1;break;case B.Discovery:Ce[0]+=1;break;case B.Navigator:we[0]+=1;break;case B.Scarlet:Te[0]+=1;break;default:break}switch(e.line){case z.Disney:Y+=1;break;case z.NG:De+=1;break;case z.Princess:Oe+=1,X[0]+=e.concierge?2:1,X[1]+=e.nights;break;case z.RC:ke+=1,Ae+=e.nights*(e.concierge?2:1);break;case z.Virgin:Z+=1;break;default:break}switch(e.region){case V.Caribbean:K[0]+=1;break;case V.Europe:K[1]+=1;break;case V.Pacific:K[2]+=1;break;case V.West:K[3]+=1;break;default:break}});var je=Ee,Me=[u[100],l[400],le[500],ue[900],l[100]],Ne=[`Disney 🛳️`,`Virgin 🛳️`,`Princess 🌙`,`Royal 🌙`],Pe=[{type:`column`,name:`None`,data:[Math.floor(U[0].num/Y*100),Math.floor(Z/Se[1].num*100),Math.floor(W[0].nights/W[1].nights*100),Math.floor(G[0].nights/G[1].nights*100)]},{type:`column`,name:`Silver`,data:[Math.floor((U[1].num-U[0].num)/Y*100),0,0,0]},{type:`column`,name:`Gold`,data:[Math.floor((U[2].num-U[1].num)/Y*100),0,Math.floor((X[1]-W[0].nights)/W[1].nights*100),Math.floor((Ae-G[0].nights)/G[1].nights*100)]},{type:`column`,name:`Platinum`,data:[Math.floor((U[3].num-U[2].num)/Y*100),0,0,0]},{type:`column`,name:`Pearl`,data:[Math.floor((Y-U[3].num)/Y*100),0,0,0]}],Fe={nodes:[{id:V.Caribbean,color:f[800]},{id:V.Europe,color:`white`},{id:V.Pacific,color:p[600]},{id:V.West,color:u[600]},{id:`🛳`,color:ce[400]},{id:z.Disney,color:f[800]},{id:`NatGeo`,color:p[600]},{id:z.Princess,color:d[500]},{id:`Royal C`,color:ce[400]},{id:`Virgin`,color:u[900]},{id:B.Magic,color:f[900]},{id:B.Wonder,color:`white`},{id:B.Dream,color:p[600]},{id:B.Fantasy,color:u[600]},{id:B.Wish,color:`black`},{id:B.Treasure,color:f[900]},{id:B.Destiny,color:`white`},{id:B.IcelandAir,color:p[600]},{id:B.Endurance,color:p[600]},{id:B.Discovery,color:u[600]},{id:`Navigator`,color:`black`},{id:B.Scarlet,color:f[900]}],data:[[V.Caribbean,`🛳`,K[0]],[V.Europe,`🛳`,K[1]],[V.Pacific,`🛳`,K[2]],[V.West,`🛳`,K[3]],[`🛳`,z.Disney,Y],[`🛳`,`NatGeo`,De],[`🛳`,z.Princess,Oe],[`🛳`,`Royal C`,ke],[`🛳`,`Virgin`,Z],[z.Disney,B.Magic,q[0]],[z.Disney,B.Wonder,q[1]],[z.Disney,B.Dream,q[2]],[z.Disney,B.Fantasy,q[3]],[z.Disney,B.Wish,q[4]],[z.Disney,B.Treasure,q[5]],[z.Disney,B.Destiny,q[6]],[`NatGeo`,B.IcelandAir,J[0]],[`NatGeo`,B.Endurance,J[1]],[z.Princess,B.Discovery,Ce[0]],[`Royal C`,`Navigator`,we[0]],[`Virgin`,B.Scarlet,Te[0]]]},Q={padding:5,textAlign:`center`,whiteSpace:`normal`,overflow:`visible`},Ie=(0,w.memo)(()=>(0,A.jsxs)(g,{"aria-label":`cruises I have been on`,children:[(0,A.jsx)(de,{children:(0,A.jsxs)(y,{children:[(0,A.jsx)(_,{style:Q,children:`Ship 🚢`}),(0,A.jsx)(_,{style:Q,children:`Destination 📍`}),(0,A.jsxs)(_,{style:Q,children:[`Nights (`,je,`\xA0🌙)`]}),(0,A.jsx)(_,{style:Q,children:`1st 🥇`}),(0,A.jsx)(_,{style:Q,children:`Month 🗓`})]})}),(0,A.jsx)(v,{children:H.map((e,t)=>(0,A.jsxs)(w.Fragment,{children:[(t===0||e.departure.year>H[t-1].departure.year)&&(0,A.jsx)(y,{sx:{borderTop:`2px solid ${l[400]}`},children:(0,A.jsx)(_,{style:Q,colSpan:5,children:(0,A.jsx)(h,{variant:`h6`,children:e.departure.year})})}),(0,A.jsxs)(y,{children:[(0,A.jsxs)(_,{style:Q,component:`th`,scope:`row`,children:[e.line,` `,e.ship]}),(0,A.jsx)(_,{style:Q,children:e.name}),(0,A.jsx)(_,{style:Q,children:e.nights}),(0,A.jsx)(_,{style:Q,children:e.concierge?`🥇`:``}),(0,A.jsxs)(_,{style:Q,title:e.departure.format(`MMMM`),children:[e.departure.format(`MM`),` `,e.departure.format(`M`)]})]})]},`cruise-tr-${t}`))})]}));Ie.displayName=`CruiseTable`;var Le=e(me(),1);pe(),he(),re(Le.default);var $=Le.default,Re={chart:{type:`sankey`,height:800,backgroundColor:`transparent`},plotOptions:{sankey:{nodePadding:16,nodeWidth:68}}},ze=(0,w.memo)(()=>{let e=b(x).mode===`light`?`black`:`white`;return(0,A.jsx)(`figure`,{style:{margin:0,width:`100%`},children:(0,A.jsxs)(r,{highcharts:$,options:Re,children:[(0,A.jsx)(o,{enabled:!0,point:{valueDescriptionFormat:`{point.weight} {point.from} cruises have been on {point.to}.`}}),(0,A.jsx)(a,{enabled:!1}),(0,A.jsx)(s,{style:{color:e},children:`Cruises`}),(0,A.jsx)(i,{options:{name:`Cruises`,keys:[`from`,`to`,`weight`],nodes:Fe.nodes},data:Fe.data,type:`sankey`})]})})});ze.displayName=`CruiseCharts`;var Be={colors:Me,pane:{size:`100%`,innerSize:`20%`,endAngle:330},chart:{type:`column`,inverted:!0,polar:!0,backgroundColor:`transparent`,style:{marginLeft:`auto`,marginRight:`auto`}},plotOptions:{column:{stacking:`normal`,borderWidth:0,pointPadding:0,groupPadding:.15,borderRadius:`50%`}}},Ve=(0,w.memo)(()=>{let e=b(x).mode===`light`?`black`:`white`;return(0,A.jsx)(`figure`,{style:{margin:0,width:`100%`},children:(0,A.jsxs)(r,{highcharts:$,options:Be,children:[(0,A.jsx)(o,{enabled:!0}),(0,A.jsx)(a,{enabled:!1}),(0,A.jsx)(ie,{itemStyle:{color:e}}),(0,A.jsx)(n,{valueSuffix:`%`}),(0,A.jsx)(s,{style:{color:e},children:`Cruise Loyalty`}),(0,A.jsx)(ne,{labels:{align:`right`,step:1,y:3,style:{color:e}},gridLineWidth:0,categories:Ne}),(0,A.jsx)(t,{lineWidth:0,reversedStacks:!1,gridLineWidth:0,labels:{enabled:!1}}),Pe.map(e=>(0,A.jsx)(i,{...e,options:{name:e.name}},e.name))]})})});Ve.displayName=`LoyaltyCharts`;var He={chart:{type:`areaspline`,backgroundColor:`transparent`},plotOptions:{areaspline:{stacking:`normal`}}},Ue=(0,w.memo)(()=>{let e=b(x).mode===`light`?`black`:`white`;return(0,A.jsx)(`figure`,{style:{margin:0,width:`100%`},children:(0,A.jsxs)(r,{highcharts:$,options:He,children:[(0,A.jsx)(o,{enabled:!0}),(0,A.jsx)(a,{enabled:!1}),(0,A.jsx)(ie,{enabled:!1}),(0,A.jsx)(n,{shared:!0,headerFormat:`<b>{point.key}</b><br>`}),(0,A.jsx)(s,{style:{color:e},children:`Travel Days`}),(0,A.jsx)(ee,{series:{pointStart:2020}}),(0,A.jsx)(ne,{labels:{style:{color:e}}}),(0,A.jsx)(t,{labels:{style:{color:e}},title:{style:{color:e}},children:`Days`}),(0,A.jsx)(i,{type:`areaspline`,options:{name:`Vacation`},data:ve}),(0,A.jsx)(i,{type:`areaspline`,options:{name:`Work`},data:ye})]})})});Ue.displayName=`TravelDaysGraph`;var We=(0,w.memo)(()=>(0,A.jsxs)(A.Fragment,{children:[(0,A.jsx)(h,{variant:`h2`,component:`h1`,children:`Travel`}),(0,A.jsxs)(m,{container:!0,spacing:2,children:[(0,A.jsx)(m,{size:{xs:12,md:6,xxl:4},children:(0,A.jsxs)(S,{title:`Travel Map`,children:[(0,A.jsx)(R,{}),(0,A.jsx)(F,{}),(0,A.jsx)(Ue,{})]})}),(0,A.jsx)(m,{size:{xs:12,md:6,xxl:4},children:(0,A.jsxs)(S,{title:`Cruise Charts`,children:[(0,A.jsx)(ze,{}),(0,A.jsx)(Ve,{})]})}),(0,A.jsx)(m,{size:{xs:12,xxl:4},children:(0,A.jsx)(S,{title:`Cruises`,children:(0,A.jsx)(Ie,{})})})]})]}));We.displayName=`TravelMap`;export{We as default};
-//# sourceMappingURL=travel-map-CV-__66Y.js.map
