+++ docs/index.html
@@ -62,7 +62,7 @@
      /  /__/  / / /_/ / / ___/  /
     /________/_/\__  /_/\___/__/
                /____/           -->
-    <script type="module" crossorigin src="/assets/index-DBeL36ud.js"></script>
+    <script type="module" crossorigin src="/assets/index-BeqMPe7A.js"></script>
     <link rel="modulepreload" crossorigin href="/assets/rolldown-runtime-DF2fYuay.js">
     <link rel="modulepreload" crossorigin href="/assets/charts-DOPwoKjG.js">
     <link rel="modulepreload" crossorigin href="/assets/react-vendor-HqBF-KSi.js">
